requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
transformers==4.35.2
opencv-python==4.8.1.78
Pillow==10.1.0
PyPDF2==3.0.1
pdfplumber==0.10.2
pytesseract==0.3.10
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
kaggle==1.5.16
requests==2.31.0
python-dotenv==1.0.0
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
wordcloud==1.9.2
